+++ docs/HISTORY.rst
@@ -8,6 +8,8 @@ Changelog
 - Add start of membership to PDF.
   [msom]
 
+- Remove more unprocessable tags in organization portrait for PDF generation.
+  [msom]
 
 0.7.1 (2015-04-08)
 ~~~~~~~~~~~~~~~~~~